210104 | 시간에 따른 기상정보 api 요청에 따른 결과 확인|UTC값을 현재 시간으로 변경 후 date에 저장

+++ app/src/main/java/com/heon9u/alarm_weather_app/Dto/Feels_like.java
@@ -1,33 +1,33 @@
 package com.heon9u.alarm_weather_app.Dto;
 
 public class Feels_like {
-    private float morn;
-    private float day;
-    private float eve;
-    private float night;
+    private double morn;
+    private double day;
+    private double eve;
+    private double night;
 
-    public float getMorn() {
+    public double getMorn() {
         return morn;
     }
-    public void setMorn(float morn) {
+    public void setMorn(double morn) {
         this.morn = morn;
     }
-    public float getDay() {
+    public double getDay() {
         return day;
     }
-    public void setDay(float day) {
+    public void setDay(double day) {
         this.day = day;
     }
-    public float getEve() {
+    public double getEve() {
         return eve;
     }
-    public void setEve(float eve) {
+    public void setEve(double eve) {
         this.eve = eve;
     }
-    public float getNight() {
+    public double getNight() {
         return night;
     }
-    public void setNight(float night) {
+    public void setNight(double night) {
         this.night = night;
     }
